# server/app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Tiquet API"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/tasks/{id}": {
      "get": {
        "description": "Returns a task",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Task id",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Tasks"
        ],
        "responses": {
          "200": {
            "description": "Returns a task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "403": {
            "description": "You can't perform this action"
          }
        }
      },
      "delete": {
        "description": "Deletes a task",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Task id",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Tasks"
        ],
        "responses": {
          "200": {
            "description": "Task deleted"
          },
          "403": {
            "description": "You can't perform this action"
          }
        }
      }
    },
    "/tasks/{id}/list": {
      "patch": {
        "description": "Move a task to another list",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Task id",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "listId": {
                    "type": "number",
                    "description": "Destination list"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ],
        "responses": {
          "200": {
            "description": "Task updated"
          },
          "400": {
            "description": "Missing params"
          },
          "403": {
            "description": "You can't perform this action"
          }
        }
      }
    },
    "/tasks/{id}/description": {
      "patch": {
        "description": "Update task's description",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Task id",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "description": "New description"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ],
        "responses": {
          "200": {
            "description": "Task updated"
          },
          "400": {
            "description": "Missing params || New description can't be the same as the old one."
          },
          "403": {
            "description": "You can't perform this action"
          }
        }
      }
    },
    "/tasks/{id}/title": {
      "patch": {
        "description": "Update task's title",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Task id",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "description": "New title"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ],
        "responses": {
          "200": {
            "description": "Task updated"
          },
          "400": {
            "description": "Missing params || New title can't be the same as the old one."
          },
          "403": {
            "description": "You can't perform this action"
          }
        }
      }
    },
    "/tasks/{id}/priority": {
      "patch": {
        "description": "Update task's priority",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Task id",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "number",
                    "description": "New priority id"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ],
        "responses": {
          "200": {
            "description": "Task updated"
          },
          "400": {
            "description": "Missing params || New priority can't be the same as the old one."
          },
          "403": {
            "description": "You can't perform this action"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Priority": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "user_id": {
            "type": "number"
          },
          "list_id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "uid": {
            "type": "string"
          },
          "createdat": {
            "type": "number"
          },
          "description": {
            "type": "number"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          }
        }
      },
...
    }
  },
...
}